==> 1-prelim-code.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

### Loading up

train_features_csv = "./data/dengue_features_train.csv"
train_labels_csv = "./data/dengue_labels_train.csv"

features = pd.read_csv(train_features_csv)
labels = pd.read_csv(train_labels_csv)

features.head(5)

labels.head()

# adding the label to features and renaming to "train" for simplicity
features["total_cases"] = labels.total_cases

train = features

### First looks

train.head(10)

train.info()

# train.dropna(axis=0, inplace=True)

print(train.total_cases.describe())
train[train["total_cases"]==0].count()["total_cases"]

# no missing targets (which is expected I suppose). Nothing crazy really, 100 "zero" cases sounds right, and I don't
# think you can have imbalanced classes in regression?

train.to_csv("./data/train.csv")
