==> 4-null-handling.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

from sklearn.preprocessing import PolynomialFeatures, StandardScaler
from sklearn.linear_model import Lasso
from sklearn.model_selection import train_test_split, cross_val_score

from keras.models import Sequential
from keras.layers import Dense, Dropout
from keras import regularizers
from keras.callbacks import EarlyStopping

Notes:
- At first I went through and changed all null values of the average of all entries for the same week, but that actually made scores worse. So now I'm going to drop all missing ndvi_ne data, input missing data as the average (it should be a lot less imputed data), and retry
- The second method was much better on the train/test split (~5 MAE points). My next step will be to separate San Juan from Iquitos, because there's probably so little geographic relation between the two that there shouldnt be any influence of one on the other

train_csv = "./data/train.csv"
train = pd.read_csv(train_csv)

 train.drop("Unnamed: 0", axis=1, inplace=True)

train.info()

pd.set_option('display.max_rows', 1500)
pd.set_option('display.max_columns', 30)

plan: try to set missing data as an average of all other matching weeks of the year

train.head(10)

train.info()

# a function/loop that can take null values and replace them with column averages

# testing out a few code snippets
train.iloc[10]["weekofyear"]

# more testing
(train[train["weekofyear"]==10].precipitation_amt_mm).mean()

# more testing
pd.isnull(train["ndvi_ne"][9])

train.dropna(axis=0, subset=["ndvi_ne"], inplace=True)

train.reset_index(drop=True, inplace=True)

# checking unique values beforehand
for i in train.columns:
    print(i,":", len(train[i].unique()))

for column_name in train.columns:
    for index, entry in enumerate(train[column_name]):
        if column_name not in ["city", "week_start_date"]:
            if pd.isnull(entry):
                train.loc[index,column_name] = (train[
                                                    train["weekofyear"]==train.iloc[index]["weekofyear"]
                                                ][column_name]).mean()

# this loop is essentially zooming in on a particular NaN entry, taking the index, using the index to find the 
# matching week of the year, and then inputting the average of that column for all entries with the same week of 
# the year in place of the NaN value

train.info()

# checking unique values afterwards
for i in train.columns:
    print(i,":", len(train[i].unique()))

# not so many new values, good

test_csv = "./data/dengue_features_test.csv"
test = pd.read_csv(test_csv)

test.info()

for column_name in test.columns:
    for index, entry in enumerate(test[column_name]):
        if column_name not in ["city", "week_start_date"]:
            if pd.isnull(entry):
                test.loc[index,column_name] = (test[
                                                    test["weekofyear"]==test.iloc[index]["weekofyear"]
                                                ][column_name]).mean()

test.info()

test[test["ndvi_ne"].isnull()]
# look at that...

train[train["weekofyear"]==1]

train[train["weekofyear"]==53]
# filling these in probably isn't good practice so i'll ditch them

train.dropna(inplace=True)

train.info()

train.to_csv("./data/cleaned_train.csv")

# i can't drop the test entry so i'll have to fill it in

test[test["weekofyear"]==53]

test.loc[87, "weekofyear"] = 1

for column_name in test.columns:
    for index, entry in enumerate(test[column_name]):
        if column_name not in ["city", "week_start_date"]:
            if pd.isnull(entry):
                test.loc[index,column_name] = (test[
                                                    test["weekofyear"]==test.iloc[index]["weekofyear"]
                                                ][column_name]).mean()

test.info()
# cool

test.to_csv("./data/test.csv")

### Separating cities
