==> 6-city-separation.ipynb <==
### Separating SJ from IQ

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

train_csv = "./data/cleaned_train.csv"
train = pd.read_csv(train_csv)

train.head()

train.drop("Unnamed: 0", axis=1, inplace=True)

sj = train["city"]=="sj"
iq = train["city"]=="iq"

train_sj = train[sj]
train_iq = train[iq]

train_sj.head()

train_iq.head()

train_sj.to_csv("./data/train_sj.csv")
train_iq.to_csv("./data/train_iq.csv")

# easy
