==> 8-hypertuning.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

from sklearn.preprocessing import PolynomialFeatures, StandardScaler
from sklearn.linear_model import Lasso
from sklearn.model_selection import train_test_split, cross_val_score

from keras.models import Sequential
from keras.layers import Dense, Dropout
from keras import regularizers
from keras.callbacks import EarlyStopping

!pip install talos

# I looked around online for methods on hyperparameter tuning and found one where someone built their own
# gridsearch-like package to use specifically with keras... so here we go

import talos as talos

def hypertuner(Xtrain, ytrain, Xtest, ytest, params):
    
    network = Sequential()
    
    # input layer
    network.add(Dense(units=Xtrain.shape[1],
                      activation="relu",
                      input_dim=Xtrain.shape[1],
                      kernel_regularizer=params["regularizer"](params["lambda"])))
    
    network.add(Dropout(params["Dropout"]))
    
    # first optional hidden layer
    if params["hidden_layers"] >= 1:
        network.add(Dense(units=params["hidden_nodes"],
                          activation="relu",
                          kernel_regularizer=params["regularizer"](params["lambda"])))
        
        network.add(Dropout(params["Dropout"]))
        
    # second optional hidden layer
    if params["hidden_layers"] >= 2:
        network.add(Dense(units=params["hidden_nodes"],
                          activation="relu",
                          kernel_regularizer=params["regularizer"](params["lambda"])))
        
        network.add(Dropout(params["Dropout"]))    

    # third optional hidden layer    
    if params["hidden_layers"] >= 3:
        network.add(Dense(units=params["hidden_nodes"],
                          activation="relu",
                          kernel_regularizer=params["regularizer"](params["lambda"])))
        
        network.add(Dropout(params["Dropout"]))

    # output layer
    network.add(Dense(units=1,
                      activation=None,
                      kernel_regularizer=params["regularizer"](params["lambda"])))
    
    # compiling
    network.compile(loss="mean_absolute_error", optimizer="adam")
    
    # fitting
    fitted_model = network.fit(Xtrain, 
                          ytrain, 
                          validation_data=[Xtest, ytest], 
                          epochs=params["epochs"],
                          verbose=0)
    
    return fitted_model, network

parameters = {
    "regularizer" : [regularizers.l1, regularizers.l2],
    "lambda" : [.005, .01, .05, .1],
    "Dropout" : [.1, .25, .5],
    "hidden_layers" : [0, 1, 2, 3],
    "hidden_nodes" : [50, 100, 150, 200],
    "epochs" : [300, 500, 1000]
}

train = pd.read_csv("./data/cleaned_train.csv")

train.drop(["Unnamed: 0", "week_start_date"], axis=1, inplace=True)

train = pd.get_dummies(train)

target = "total_cases"
features = [x for x in train if x != target]
y = train[target]
X = train[features]
X_train, X_test, y_train, y_test = train_test_split(X,y)

ss = StandardScaler()
X = ss.fit_transform(X)

y.values

X

search = talos.Scan(x=X,
                    y=y.values,
                    params=parameters,
                    model=hypertuner,
                    dataset_name="whatthehell")

search_1 = talos.Reporting("./train_.csv")

search_1.data.sort_values("hidden_layers",ascending=False).head(10)

##### second run through

def hypertuner(Xtrain, ytrain, Xtest, ytest, params):
    
    network = Sequential()
    
    # input layer
    network.add(Dense(units=Xtrain.shape[1],
                      activation="relu",
                      input_dim=Xtrain.shape[1],
                      kernel_regularizer=params["regularizer"]))
    
    network.add(Dropout(params["Dropout"]))
    
    # first optional hidden layer
    if params["hidden_layers"] >= 1:
        network.add(Dense(units=params["hidden_nodes"],
                          activation="relu",
                          kernel_regularizer=params["regularizer"]))
        
        network.add(Dropout(params["Dropout"]))
        
    # second optional hidden layer
    if params["hidden_layers"] >= 2:
        network.add(Dense(units=params["hidden_nodes"],
                          activation="relu",
                          kernel_regularizer=params["regularizer"]))
        
        network.add(Dropout(params["Dropout"]))    

    # output layer
    network.add(Dense(units=1,
                      activation=None,
                      kernel_regularizer=params["regularizer"]))
    
    # compiling
    network.compile(loss="mean_absolute_error", optimizer="adam")
    
    early = EarlyStopping(monitor="val_loss", min_delta=0, patience=2)
    
    # fitting
    HISTORY = network.fit(Xtrain, 
                          ytrain, 
                          validation_data=[Xtest, ytest], 
                          epochs=params["epochs"],
                          callbacks=[early],
                          verbose=0)
    
    return HISTORY, network

parameters = {
    "regularizer" : [regularizers.l2(.001), regularizers.l2(.005), regularizers.l2(.01)],
#     "lambda" : [.001, .005, .01],
    "Dropout" : [.15, .25],
    "hidden_layers" : [0, 1, 2],
    "hidden_nodes" : [75, 150, 200],
    "epochs" : [500, 1000]
#     "min_delta" : [.01, .05, .1],
#     "patience" : [5,10,20]
}

search = talos.Scan(x=X,
                    y=y.values,
                    params=parameters,
                    model=hypertuner,
                    dataset_name="with_earlystop")

search.data.sort_values("val_loss", ascending=True).head()

##### third runthrough with polynomial weather

train = pd.read_csv("./data/cleaned_train.csv")
train.drop(["Unnamed: 0", "week_start_date"], axis=1, inplace=True)
train = pd.get_dummies(train)

non_weather = "year weekofyear ndvi_ne ndvi_nw ndvi_se ndvi_sw total_cases city_sj city_iq".split()
weather_features = [x for x in train.columns if x not in non_weather]

pf = PolynomialFeatures(degree=2, include_bias=False)
pf_feat = pd.DataFrame(data=pf.fit_transform(train[weather_features]), columns=pf.get_feature_names(weather_features))

train_weather = train.copy()
train_weather.drop(labels=weather_features, axis=1, inplace=True)
train_weather = train_weather.join(pf_feat, how="outer")

target = "total_cases"
features = [x for x in train_weather if x != target]
y = train_weather[target]
X = train_weather[features]

ss = StandardScaler()
X = ss.fit_transform(X)


def hypertuner(Xtrain, ytrain, Xtest, ytest, params):
    
    network = Sequential()
    
    # input layer
    network.add(Dense(units=Xtrain.shape[1],
                      activation="relu",
                      input_dim=Xtrain.shape[1],
                      kernel_regularizer=params["regularizer"]))
    
    network.add(Dropout(params["Dropout"]))
    
    # first optional hidden layer
    if params["hidden_layers"] >= 1:
        network.add(Dense(units=params["hidden_nodes"],
                          activation="relu",
                          kernel_regularizer=params["regularizer"]))
        
        network.add(Dropout(params["Dropout"]))
        
    # second optional hidden layer
    if params["hidden_layers"] >= 2:
        network.add(Dense(units=params["hidden_nodes"],
                          activation="relu",
                          kernel_regularizer=params["regularizer"]))
        
        network.add(Dropout(params["Dropout"]))    

    # output layer
    network.add(Dense(units=1,
                      activation=None,
                      kernel_regularizer=params["regularizer"]))
    
    # compiling
    network.compile(loss="mean_absolute_error", optimizer="adam")
    
    early = EarlyStopping(monitor="val_loss", min_delta=0, patience=params["patience"])
    
    # fitting
    HISTORY = network.fit(Xtrain, 
                          ytrain, 
                          validation_data=[Xtest, ytest], 
                          epochs=params["epochs"],
                          callbacks=[early],
                          verbose=0)
    
    return HISTORY, network

parameters = {
    "regularizer" : [regularizers.l2(.001), regularizers.l2(.005), regularizers.l2(.01)],
#     "lambda" : [.001, .005, .01],
    "Dropout" : [.15, .25],
    "hidden_layers" : [0, 1, 2],
    "hidden_nodes" : [75, 150, 200],
    "epochs" : [500, 1000],
#     "min_delta" : [.01, .05, .1],
    "patience" : [3,5,10]
}

search_3 = talos.Scan(x=X,
                    y=y.values,
                    params=parameters,
                    model=hypertuner,
                    dataset_name="with_earlystop")

search_3.data.sort_values("val_loss", ascending=True).head(10)

#### fourth runthrough with whatever, none of these are scoring better than the others

def hypertuner(Xtrain, ytrain, Xtest, ytest, params):
    
    network = Sequential()
    
    # input layer
    network.add(Dense(units=Xtrain.shape[1],
                      activation="relu",
                      input_dim=Xtrain.shape[1],
                      kernel_regularizer=params["regularizer"]))
    
    network.add(Dropout(params["Dropout"]))
    
    # first optional hidden layer
    if params["hidden_layers"] >= 1:
        network.add(Dense(units=params["hidden_nodes"],
                          activation="relu",
                          kernel_regularizer=params["regularizer"]))
        
        network.add(Dropout(params["Dropout"]))
        
    # second optional hidden layer
    if params["hidden_layers"] >= 2:
        network.add(Dense(units=params["hidden_nodes"],
                          activation="relu",
                          kernel_regularizer=params["regularizer"]))
        
        network.add(Dropout(params["Dropout"]))    

    # output layer
    network.add(Dense(units=1,
                      activation=None,
                      kernel_regularizer=params["regularizer"]))
    
    # compiling
    network.compile(loss="mean_absolute_error", optimizer="adam")
    
    early = EarlyStopping(monitor="val_loss", min_delta=0, patience=params["patience"])
    
    # fitting
    HISTORY = network.fit(Xtrain, 
                          ytrain, 
                          validation_data=[Xtest, ytest], 
                          epochs=params["epochs"],
                          callbacks=[early],
                          verbose=0)
    
    return HISTORY, network

parameters = {
    "regularizer" : [regularizers.l1(.001), regularizers.l1(.005), regularizers.l1(.01)],
#     "lambda" : [.001, .005, .01],
    "Dropout" : [.10, .20, .30],
    "hidden_layers" : [0, 1, 2],
    "hidden_nodes" : [75, 150, 200],
    "epochs" : [1000],
#     "min_delta" : [.01, .05, .1],
    "patience" : [3,5,10]
}

search = talos.Scan(x=X,
                    y=y.values,
                    params=parameters,
                    model=hypertuner,
                    dataset_name="l1_early")

search.data.sort_values("val_loss", ascending=True).head(10)
